[US1A] Convert to new data format and add cooling system (#15)
* Add wanted systems to extraction function

* Add route for model generation only

* Update segment and pump mappers

Updated the mapper for FlowSegments to support new data format, and
to support rectangular segments (ducts) by calculating hydraulic
diameter. Added support for pumps with the 3 currently implemented
control types. The externally controlled pump must be connected to
controllers when controls are implemented

* Add .vscode folder to gitignore

* Update radiatore mapper

* Add defaults for internal parameters

package_name, model_name, x_pos and y_pos are set by external functions.
To remove errors in IDEs, they have been initialized with
default values.

* Add initialization of mo_file

To avoid warnings in IDEs, the mo_file variable, which is set by
external functions, has been initialized as an empty string.

* Update HeatExchanger mapper

* Add fluids package

* Updated mapping for bends

Bends are now mapped to a fixed resistance pressure drop, thus avoiding
complicated models in Modelica, that slow simulations down.

The loss coefficient is calculated through the fluids package. The
package calculates the Reynolds number and uses this for calculating
the Rennels loss coefficient.

* Updated mapper for tees

Tees now have the correct nominal flow. Calculation of nominal pressure
loss is still missing.

* Update valve mapper & add shunt mapper

Mapping of valves was updated, since valve types are now defined by the
component type. Shunts have also been added. The function for length
calculation has been split in two.

* Add reductions

* Add reductions

* Replace diameter calculation to a single function

* Change indentation

* Fix hardcoded bend radius

* Change parsing procedure

Instead of writing all components directly in the model string, a
separate component string is created and then added to the model string.
This will support a new way to handle connections, that will avoid a
second loop.

* Base components' medium on system type

Medium of the components is now based on the default medium of the
system, as defined in the script.

In the future, this should be a user input.

* Update routes

Currently, only create_modelica_model should be used!

* Add port names to all components and ports

Port names are now defined alongside the component template to ensure
that they are based on the modelica models. Port names in connections
are now based on the component port names.

* Convert object mapping to class structure

* Add UML class diagram

* Change to model creation only

* Add to gitignore

* Delete non-OOP files

* Finish class hierarchy for connection mapping

* Restructure